telebot_files/requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
configparser==5.0.1
cryptography==3.3.1
DateTime==4.3
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
idna==2.10
mysql-connector-python==8.0.22
mysqlclient==2.0.1
protobuf==3.14.0
pycodestyle==2.7.0
pycparser==2.20
PyMySQL==1.0.1
python-dotenv==0.15.0
python-telegram-bot==13.7
pytz==2020.5
requests==2.25.1
schedule==0.6.0
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.20
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
virtualenv==20.2.2
zope.interface==5.2.0
